% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.fit.R
\name{rad.fit}
\alias{rad.fit}
\title{Rank Abundance Distribution Fit (RAD)}
\usage{
rad.fit(Vdata, name = FALSE, plot = F, method = "Log10",
  mixing = FALSE, ...)
}
\arguments{
\item{Vdata}{data.frame. Abundance dataset e.g. the first column is species name then communities}

\item{name}{logial. used just plot is \code{TRUE}. \code{Default = FALSE}}

\item{plot}{logical. used to plot RAD. \code{Default = FALSE}}

\item{method}{different methods to RAD plot. e.g. 'Abundance',  'Log2',  'Log10',  'Rel.Abun'.}

\item{mixing}{Combining all abundance dataset in a plot}

\item{...}{optional}
}
\value{
If \code{plot = FALSE}. The function creates a list of all communities. Another way, 
plot RAD.
}
\description{
RAD is used to show structure community of species using different approaches to fited our date like Abundance, Relative Abundance and Log2.

These methods are the most used on ecology.
}
\details{
The function \code{rad.fit} uses three different methods for fitting a curve. \code{Abundance} is a bit used.
}
