% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.fit.R
\name{rad.fit}
\alias{rad.fit}
\title{Rango Abundance Distribution Fit}
\usage{
rad.fit(Vdata, name = FALSE, plot = F, method = "Log10",
  mixing = FALSE, ...)
}
\arguments{
\item{Vdata}{Two Community with Abundance of species.}

\item{name}{asd asd asd asd as}

\item{plot}{a das das das}

\item{method}{a dasd as das dasdsa}

\item{mixing}{asdad as da}

\item{...}{optional}
}
\value{
If all inputs are integer and logical, then the output
  will be an integer. If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{Rango Abundance Curve} returns the abundance models.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
